Created individual files for Vitals and Diagnosis class. Began updating unit tests accordingly.
