+++ test_avghr.py
@@ -1,5 +1,6 @@
 import pytest
-import HR_allfuncs as hr
+import Vitals as v
+import Diagnosis as d
 import numpy as np
 
 # create a sine wave array to test peak finder, instant heart rate function
@@ -26,7 +27,7 @@ def test_numinputs():
     Test if the number of inputs for hr_averaging() is valid, otherwise throw error
     """
     with pytest.raises(TypeError):
-        x = hr.Vitals()
+        x = v.Vitals()
         x.hr_averaging()
 
 
@@ -36,7 +37,7 @@ def test_nonzero_avgingtime():
     Test if the averaging time is a non-zero and positive number, otherwise throw error
     """
     with pytest.raises(ValueError):
-        x = hr.Vitals()
+        x = v.Vitals()
         x.hr_averaging(0, array_test_time)
 
 
@@ -46,7 +47,7 @@ def test_validlen_avgingtime():
     Test if the averaging time less than or equal to the length of ECG acquisition time, otherwise throw error
     """
     with pytest.raises(ValueError):
-        x = hr.Vitals()
+        x = v.Vitals()
         x.hr_averaging(15, array_test_time)
 
 def test_isnumber_avgingtime():
@@ -55,7 +56,7 @@ def test_isnumber_avgingtime():
     Test if the averaging time is a number, otherwise throw error
     """
     with pytest.raises(ValueError):
-        x = hr.Vitals()
+        x = v.Vitals()
         x.hr_averaging("word", array_test_time)
 
 def test_fraction_divby0():
@@ -64,7 +65,7 @@ def test_fraction_divby0():
     Test if the averaging time is a valid fraction, throw error if zero division occurs
     """
     with pytest.raises(ZeroDivisionError):
-        x = hr.Vitals()
+        x = v.Vitals()
         x.hr_averaging('1/0', array_test_time)
 
 def test_fraction_validsyntax():
@@ -73,7 +74,7 @@ def test_fraction_validsyntax():
     Test if the averaging time uses valid fraction syntax
     """
     with pytest.raises(ValueError):
-        x = hr.Vitals()
+        x = v.Vitals()
         x.hr_averaging('1/2/3', array_test_time)
 
 
@@ -82,7 +83,7 @@ def test_avghr_withfraction():
     .. function:: test_avghr_withfraction():
     Test if the avghr function calculates the correct avg HR using a fraction
     """
-    x = hr.Vitals()
+    x = v.Vitals()
     x.hr_averaging('1/4', array_test_time)
     assert x.avg_hr_val == 60
 
@@ -92,7 +93,7 @@ def test_avghr_with_float_as_string():
     .. function:: test_avghr_with_float_as_string():
     Test if the avghr function calculates the correct avg HR using a decimal value passed as a string
     """
-    x = hr.Vitals()
+    x = v.Vitals()
     x.hr_averaging('1/4', array_test_time)
     assert x.avg_hr_val == 60
 
@@ -101,7 +102,7 @@ def test_avghr_with_float():
     .. function:: test_avghr_with_float():
     Test if the avghr function calculates the correct avg HR using a decimal value
     """
-    x = hr.Vitals()
+    x = v.Vitals()
     x.hr_averaging('1/4', array_test_time)
     assert x.avg_hr_val == 60
 
@@ -111,7 +112,7 @@ def test_tachylim_valid():
     Test if the tachycardia limit is a valid threshold
     """
     with pytest.raises(ValueError):
-        x = hr.Diagnosis()
+        x = d.Diagnosis()
         x.tachy(80, -1)
 
 
@@ -121,7 +122,7 @@ def test_tachystring():
     Test if the tachy_limit input is a valid number
     """
     with pytest.raises(TypeError):
-        x = hr.Diagnosis
+        x = d.Diagnosis()
         x.tachy("word")
 
 
@@ -130,8 +131,8 @@ def test_tachy_present():
     .. function:: test_tachy_present():
     Test to see if tachycardia is present
     """
-    x = hr.Diagnosis()
-    x.tachy(200, tachy_limit=100)
+    x = d.Diagnosis()
+    x.tachy()
     assert x.tachy_result is True
 
 
@@ -140,8 +141,8 @@ def test_tachy_not_present():
     .. function:: test_tachy_not_present():
     Test to see if tachycardia is not present
     """
-    x = hr.Diagnosis()
-    x.tachy(80, tachy_limit=100)
+    x = d.Diagnosis()
+    x.tachy()
     assert x.tachy_result is False
 
 
@@ -151,8 +152,8 @@ def test_brachylim_valid():
     Test if the brachycardia limit is a valid threshold
     """
     with pytest.raises(ValueError):
-        x = hr.Diagnosis()
-        x.brachy(80, -1)
+        x = d.Diagnosis()
+        x.brachy()
 
 
 def test_brachystring():
@@ -161,8 +162,8 @@ def test_brachystring():
     Test if the brachy_limit input is a valid number
     """
     with pytest.raises(TypeError):
-        x = hr.Diagnosis()
-        x.brachy(80, "word")
+        x = d.Diagnosis()
+        x.brachy()
 
 
 def test_brachy_present():
@@ -170,8 +171,8 @@ def test_brachy_present():
     .. function:: test_brachy_present():
     Test to see if brachycardia is present
     """
-    x = hr.Diagnosis()
-    x.brachy(30, brachy_limit=60)
+    x = d.Diagnosis()
+    x.brachy()
     assert x.brachy_result is True
 
 
@@ -180,6 +181,6 @@ def test_brachy_not_present():
     .. function:: test_brachy_not_present():
     Test to see if brachycardia is not present
     """
-    x = hr.Diagnosis()
-    x.brachy(65, brachy_limit=60)
+    x = d.Diagnosis()
+    x.brachy()
     assert x.brachy_result is False
